+++ guild.go
@@ -1,4 +1,4 @@
-package main
+package goff
 
 import "database/sql"
 
@@ -10,11 +10,11 @@ type Guild struct {
 	WelcomeChannel string
 	PuzzleChannel  string
 	PuzzleRole     sql.NullString
+	Prefixes       []string
 }
 
 type GuildService interface {
 	Guild(id string) (*Guild, error)
-	Guilds() ([]*Guild, error)
 	CreateGuild(g *Guild) error
 	DeleteGuild(g *Guild) error
 	GuildUsers(g *Guild) ([]*User, error)
